percent, innerclient path

+++ src/main/java/ru/bereshs/hhworksearch/model/VacancyEntity.java
@@ -41,7 +41,7 @@ public class VacancyEntity  {
     private String skillStringList;
 
     @Transient
-    private Long percent;
+    private Integer percent;
     @Transient
     private String filterResult;
     @Override
